flatpak: add a publish_beta template job
This is not meant to be used as part of every CI pipeline, but
instead on special occasions like when tagging a beta release.

Ideally there are special requirements for pushing to flathub-beta
that you need to make sure you follow if you want to use this.
Whats needed is the following:

* Ask a sysadmin to add a FLATHUB_BETA_TOKEN to your project and
  make sure your branch is protected
* Make sure you application depends on a runtime found in Flathub
  or Flathub beta. You shouldn't use this with the Nightly runtime.
* Make sure your 'branch' in the Flatpak manifest is beta
* Change only: and when: yaml keys to fit your needs

Close #6