templates: Skip gnomeos-test-sysext job if no credentials
We can't demand non-member contributors to setup openQA credentials
on their forks. So, let's run this job only in pipelines with these
credentials.

The proposed workflow is:

A member sends a MR and, by our current policy, the job is run
because these credentials are available in the parent project.

A non-member sends a MR and, by our current policy, the job does not
run until the reviewer assigns the MR to marge-bot, and then it runs.

See https://gitlab.gnome.org/GNOME/openqa-tests
See https://gitlab.gnome.org/GNOME/gnome-build-meta/-/issues/837