gnomeos-basic-ci: Add grcov job for C projects
grcov is a code coverage tool, primarily for Rust projects.

Even though grcov is primarily used for Rust projects,
it can also be used instead of lcov/gcov as it produces
better and more accessible outputs.

In this case we will only set it up for C projects
using only meson. Rust projects are slightly more
complicated and to be done later.

The job can be disabled with `grcov-c-eanbled: false`
input.