release-service: Automatically discover and upload tarballs
Previously, we required the tarball to be passed by absolute
path to the component, which has proved error prone as
both the name and path variables need to be global and known
ahead of time.

For basic-release-with-dist we implemented a different scheme
outlined in GNOME/citemplates#23 and it seems to work well
so far.

Implement the same logic for the release service. Change
the tarball-artifact-path value to "public-dist". If it's
changed, assume we are getting passed an absolute path
to the tarball, otherwise look for tarballs to upload
from the directory.

Close GNOME/citemplates#23