gnomeos-basic-ci: Add sanitizer jobs
Add jobs that compile and runs the testsuite
with all the usual sanitizers we use.

All of them have a $sanitizer-enabled boolean
input which defaults to true so that it works ootb
but it can be disabled when needed.

Additionally extra options can be passed using the
usual sanitizer _OPTIONS variables