flatpak: Include the default-rules template by default
Till now we'd leave it to projects explicitly including
the template, however a lot of projects were setup
years ago and only include the flatpak template.

We can include the rules by default to avoid having
to update all the projects and have a simpler setup.

This could cause issues as rules may be overridden
by projects defining their own (but that's fine) but
more importantly there's an edge case where people
might first define their default/workflow and put their
include: afterwards in which cause it will silently,
override them as the last statement wins. However
that's very unlikely as in the majority of the projects
we define the includes at the very top of the file.