flatpak: Add a job that will fail if REPO_TOKEN is still set
In e2450cf886da6f01f1657c3db53d2066c276faaf we changed the
way we pass the token to flat-manager to use a File Variable.

Now that almost everyone has migrated, add a job that will fail
if the variable is still set to make sure we catch the last ones
as well.