flatpak: Switch to quay.io as the OCI registry
The packages that GHCR creates on push are private
by default and require authentication to be pulled.

(Unless they are pushed through a GH Action)

Surprisingly there is no REST API to change the visibility
of a package, but only to get/delete/restore a package,
which means we'd have to change the visibility of each ref
manually.

Switch to quay.io since we already host other infra images
there anyway.

Additionally rename the env vars we use to be more generic