flatpak: Always set the default-branch
Till now we were using the default name, master, unless
exlicitly specified otherwise.

We have to keep building and pushing the master as its
the default branch in the Nightly repository, but
we can override the branch we build for Merge Requests
and test builds.

This way Bundles will always be named after the MR
number they are associated with, and you can additionally
install test builds without replacing the Nightly
installation.